Refactor: move all the spec related code to one file :recycle:
