+++ src/cfnlint/rules/resources/iam/Permissions.py
@@ -4,7 +4,8 @@ SPDX-License-Identifier: MIT-0
 """
 import json
 import six
-from cfnlint.helpers import convert_dict, load_resource
+from cfnlint.helpers import convert_dict
+from cfnlint.specs import load_resource
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
 from cfnlint.data import AdditionalSpecs
