+++ src/cfnlint/rules/resources/lmbd/DeprecatedRuntime.py
@@ -3,7 +3,7 @@ Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
 from datetime import datetime
-from cfnlint.helpers import load_resource
+from cfnlint.specs import load_resource
 from cfnlint.data import AdditionalSpecs
 from cfnlint.rules import CloudFormationLintRule
