+++ src/cfnlint/rules/resources/properties/AllowedPattern.py
@@ -6,7 +6,7 @@ import re
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
 
-from cfnlint.helpers import RESOURCE_SPECS
+from cfnlint.specs import RESOURCE_SPECS
 
 
 class AllowedPattern(CloudFormationLintRule):
