+++ src/cfnlint/rules/resources/properties/AllowedValue.py
@@ -6,7 +6,7 @@ import json
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
 
-from cfnlint.helpers import RESOURCE_SPECS
+from cfnlint.specs import RESOURCE_SPECS
 
 
 class AllowedValue(CloudFormationLintRule):
