+++ src/cfnlint/rules/resources/properties/JsonSize.py
@@ -9,7 +9,7 @@ import six
 import cfnlint.helpers
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
-from cfnlint.helpers import RESOURCE_SPECS
+from cfnlint.specs import RESOURCE_SPECS
 
 
 class JsonSize(CloudFormationLintRule):
