+++ src/cfnlint/rules/resources/properties/ListSize.py
@@ -5,7 +5,7 @@ SPDX-License-Identifier: MIT-0
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
 
-from cfnlint.helpers import RESOURCE_SPECS
+from cfnlint.specs import RESOURCE_SPECS
 
 
 class ListSize(CloudFormationLintRule):
