+++ src/cfnlint/rules/resources/properties/NumberSize.py
@@ -7,7 +7,7 @@ import six
 from cfnlint.rules import CloudFormationLintRule
 from cfnlint.rules import RuleMatch
 
-from cfnlint.helpers import RESOURCE_SPECS
+from cfnlint.specs import RESOURCE_SPECS
 
 
 class NumberSize(CloudFormationLintRule):
