+++ src/cfnlint/transform.py
@@ -10,7 +10,8 @@ from samtranslator.parser import parser
 from samtranslator.translator.translator import Translator
 from samtranslator.public.exceptions import InvalidDocumentException
 
-from cfnlint.helpers import load_resource, convert_dict, format_json_string
+from cfnlint.helpers import convert_dict, format_json_string
+from cfnlint.specs import load_resource
 from cfnlint.data import Serverless
 from cfnlint.rules import Match, TransformError
 LOGGER = logging.getLogger('cfnlint')
