+++ test/unit/module/specs/test_complete.py
@@ -3,6 +3,9 @@ Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
 import json
+
+import cfnlint.specs
+
 from test.testlib.testcase import BaseTestCase
 from cfnlint import Runner
 from cfnlint.rules import RulesCollection
@@ -23,7 +26,7 @@ class TestComplete(BaseTestCase):
     def tearDown(self):
         """Tear Down"""
         # Reset the Spec override to prevent other tests to fail
-        cfnlint.helpers.initialize_specs()
+        cfnlint.specs.initialize_specs()
 
     def test_success_run(self):
         """Success test"""
@@ -32,7 +35,7 @@ class TestComplete(BaseTestCase):
         with open('test/fixtures/templates/override_spec/complete.json') as fp:
             custom_spec = json.load(fp)
 
-        cfnlint.helpers.set_specs(custom_spec)
+        cfnlint.specs.set_specs(custom_spec)
 
         good_runner = Runner(self.collection, filename, template, ['us-east-1'], [])
         self.assertEqual([], good_runner.run())
@@ -44,7 +47,7 @@ class TestComplete(BaseTestCase):
 
         with open('test/fixtures/templates/override_spec/complete.json') as fp:
             custom_spec = json.load(fp)
-        cfnlint.helpers.set_specs(custom_spec)
+        cfnlint.specs.set_specs(custom_spec)
 
         bad_runner = Runner(self.collection, filename, template, ['us-east-1'], [])
         errs = bad_runner.run()
