+++ test/unit/module/specs/test_update_resource_specs.py
@@ -4,6 +4,9 @@ SPDX-License-Identifier: MIT-0
 """
 import sys
 import logging
+
+import cfnlint.specs
+
 from test.testlib.testcase import BaseTestCase
 from mock import patch, MagicMock
 import cfnlint.maintenance
@@ -14,11 +17,11 @@ LOGGER.addHandler(logging.NullHandler())
 
 class TestUpdateResourceSpecs(BaseTestCase):
     """Used for Testing Resource Specs"""
-    @patch('cfnlint.maintenance.url_has_newer_version')
-    @patch('cfnlint.maintenance.get_url_content')
+    @patch('cfnlint.specs.url_has_newer_version')
+    @patch('cfnlint.specs.get_url_content')
     @patch('cfnlint.maintenance.json.dump')
-    @patch('cfnlint.maintenance.patch_spec')
-    @patch('cfnlint.maintenance.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
+    @patch('cfnlint.specs.patch_spec')
+    @patch('cfnlint.specs.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
     def test_update_resource_spec(self, mock_patch_spec, mock_json_dump, mock_content, mock_url_newer_version):
         """Success update resource spec"""
 
@@ -46,7 +49,7 @@ class TestUpdateResourceSpecs(BaseTestCase):
             builtin_module_name = '__builtin__'
 
         with patch('{}.open'.format(builtin_module_name)) as mock_builtin_open:
-            cfnlint.maintenance.update_resource_spec('us-east-1', 'http://foo.badurl')
+            cfnlint.specs.update_resource_spec('us-east-1', 'http://foo.badurl')
             mock_json_dump.assert_called_with(
                 {
                     'PropertyTypes': {},
@@ -63,42 +66,42 @@ class TestUpdateResourceSpecs(BaseTestCase):
                 sort_keys=True
             )
 
-    @patch('cfnlint.maintenance.url_has_newer_version')
-    @patch('cfnlint.maintenance.get_url_content')
+    @patch('cfnlint.specs.url_has_newer_version')
+    @patch('cfnlint.specs.get_url_content')
     @patch('cfnlint.maintenance.json.dump')
-    @patch('cfnlint.maintenance.patch_spec')
-    @patch('cfnlint.maintenance.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
+    @patch('cfnlint.specs.patch_spec')
+    @patch('cfnlint.specs.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
     def test_do_not_update_resource_spec(self, mock_patch_spec, mock_json_dump, mock_content, mock_url_newer_version):
         """Success update resource spec"""
 
         mock_url_newer_version.return_value = False
 
-        result = cfnlint.maintenance.update_resource_spec('us-east-1', 'http://foo.badurl')
+        result = cfnlint.specs.update_resource_spec('us-east-1', 'http://foo.badurl')
         self.assertIsNone(result)
         mock_content.assert_not_called()
         mock_patch_spec.assert_not_called()
         mock_json_dump.assert_not_called()
 
-    @patch('cfnlint.maintenance.multiprocessing.Pool')
-    @patch('cfnlint.maintenance.update_resource_spec')
-    @patch('cfnlint.maintenance.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
+    @patch('cfnlint.specs.multiprocessing.Pool')
+    @patch('cfnlint.specs.update_resource_spec')
+    @patch('cfnlint.specs.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
     def test_update_resource_specs_python_3(self, mock_update_resource_spec, mock_pool):
 
         fake_pool = MagicMock()
         mock_pool.return_value.__enter__.return_value = fake_pool
 
-        cfnlint.maintenance.update_resource_specs()
+        cfnlint.specs.update_resource_specs()
 
         fake_pool.starmap.assert_called_once()
 
-    @patch('cfnlint.maintenance.multiprocessing.Pool')
-    @patch('cfnlint.maintenance.update_resource_spec')
-    @patch('cfnlint.maintenance.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
+    @patch('cfnlint.specs.multiprocessing.Pool')
+    @patch('cfnlint.specs.update_resource_spec')
+    @patch('cfnlint.specs.SPEC_REGIONS', {'us-east-1': 'http://foo.badurl'})
     def test_update_resource_specs_python_2(self, mock_update_resource_spec, mock_pool):
 
         fake_pool = MagicMock()
         mock_pool.return_value.__enter__.return_value = AttributeError('foobar')
 
-        cfnlint.maintenance.update_resource_specs()
+        cfnlint.specs.update_resource_specs()
 
         mock_update_resource_spec.assert_called_once_with('us-east-1', 'http://foo.badurl')
